Les utilisateurs ont leur domaine maintenant. Sortie des users fake en dehors des classes Model. Ajout des fichiers de langues rapide.
