+++ api/src/Languages/fr/api.ts
@@ -1,0 +1,8 @@
+
+const language_api = {
+    departConfigServer: 'Départ de la configuration du serveur pour l\'API',
+    configEndedServerListening: 'Configuration terminée, départ de l\'écoute sur le serveur',
+    connectToBd: 'Connexion à la base de données ...',
+}
+
+export default language_api;
