+++ api/src/Languages/fr/errors.ts
@@ -1,0 +1,6 @@
+
+const language_error = {
+    'cantConnectToDB': 'Connexion à la base de données impossible.',
+}
+
+export default language_error;
