+++ api/src/Users/Controllers/UserController.ts
@@ -1,0 +1,5 @@
+
+
+export default class UserController {
+
+}
