+++ api/src/Users/Models/UserModel.ts
@@ -2,25 +2,12 @@
 // Database connection.
 //temp fake to have the JWT token first,
 import * as mongoDB from "mongodb";
-import {UserCredential} from "../Controllers/AuthenficationController";
-
-export const fakeUsers = [
-    {
-        username: 'datageek',
-        name: 'Hydile Durocher',
-        email: 'datageek@test.com',
-        password: '1234',
-        role: 'admin'
-    },{
-        username: 'annamontana',
-        name: "Chantal Carpediem",
-        email: 'annamontana@test.com',
-        password: 'password123member',
-        role: 'member'
-    }
-];
+import {UserCredential} from "../../Authentification/Controllers/AuthenficationController";
 
 
+/**
+ *
+ */
 export interface User {
     collection: mongoDB.Collection | null;
     username:string;
@@ -30,6 +17,9 @@ export interface User {
     role: string;
 }
 
+/**
+ *
+ */
 export default class UserModel {
     static collection: mongoDB.Collection | null;
 
@@ -66,19 +56,3 @@ export default class UserModel {
         return null;
     }
 }
-
-/*export default class UserModel {
-    username: string;
-    password: string;
-    email: string;
-    name: string;
-    role: string;
-
-    constructor(document:object) {
-        this.username = document.username;
-        this.password = document.password;
-        this.email = document.email;
-        this.name = document.name;
-        this.role = document.role;
-    }
-}*/
