+++ api/src/Users/fakeUsers.ts
@@ -1,0 +1,16 @@
+
+export const fakeUsers = [
+    {
+        username: 'datageek',
+        name: 'Hydile Durocher',
+        email: 'datageek@test.com',
+        password: '1234',
+        role: 'admin'
+    },{
+        username: 'annamontana',
+        name: "Chantal Carpediem",
+        email: 'annamontana@test.com',
+        password: 'password123member',
+        role: 'member'
+    }
+];
