+++ api/src/routes.ts
@@ -1,5 +1,4 @@
 import express from "express";
-//import HealthCheckRouter from "./Healthcheck/Routes/HealthCheckRoutes";
 
 const ApiRouter = express.Router();
